M6: move setup prompts and ProjectInfo probe into cli/skills
Moves `cli/ai/setup-prompts/` and the setup-only `cli/ai/utils/*` into
`cli/skills/content/{setup-prompts,setup-utils}` so the eval harness and
`storybook ai setup` share the same pure prompt builders the upcoming
`skills get setup` command will use, without `cli/skills` depending on
`cli/ai`.

Extracts the probing try/catch from `aiSetup` into `cli/skills/project-info.ts`
as `getProjectInfo`, returning a `ProjectInfoResult` instead of logging and
returning early; `aiSetup` now renders that result itself, preserving the
original logger.error/logger.log split. `ProjectInfo` moves with it (re-exported
from `cli/ai/types.ts` for `cli/ai/mcp` compat); `SetupInstructionsContext` moves
into `content/setup-prompts/types.ts`, scoped to the prompts that use it.

`content/setup-prompts` is bundled into the browser-platform `storybook/internal/skills`
entry, so `monorepoRule` can no longer call the fs-based `getMonorepoType()` directly —
it now reads `projectInfo.monorepoType`, computed once in `getProjectInfo`.

Updates the eval harness's import path and docs accordingly; `EVAL_SETUP_PROMPT`
behavior is unchanged.
