+++ code/core/src/cli/ai/types.ts
@@ -1,6 +1,3 @@
-import type { JsPackageManager } from 'storybook/internal/common';
-import type { SupportedRenderer } from 'storybook/internal/types';
-
 export interface AiSetupOptions {
   /** Location of the Storybook configuration directory. */
   configDir?: string;
@@ -18,37 +15,6 @@ export interface AiSetupOptions {
   runId: string;
 }
 
-export interface ProjectInfo {
-  storybookVersion: string | undefined;
-  majorVersion: number | undefined;
-  framework: string | null;
-  /** The full renderer package name, e.g. "@storybook/react" */
-  rendererPackage: string | null;
-  /** The short renderer name for docs URLs, e.g. "react" */
-  renderer?: SupportedRenderer;
-  builderPackage: string | null;
-  addons: string[];
-  configDir: string;
-  storiesPaths: string[];
-  /** Whether the project uses TypeScript ('ts') or JavaScript ('js'), inferred from the main config file extension. */
-  language: 'ts' | 'js';
-  /** Detected package manager (npm, yarn, pnpm, bun), if known. */
-  packageManager: JsPackageManager;
-  /** Pretty name of the detected package manager, if known. */
-  packageManagerName?: string;
-  /** Whether the project's preview file uses the CSF Factory format. */
-  hasCsfFactoryPreview: boolean;
-  /** Whether the user has requested to be onboarded into Storybook. */
-  needsUserOnboarding: boolean;
-}
-
-export interface SetupInstructionsContext {
-  configDir: string;
-  docsUrl: (path: string) => string;
-  mswInstall: string;
-  needsUserOnboarding: boolean;
-  packageManager: JsPackageManager;
-  packageManagerName: string | undefined;
-  tsx: string;
-  ts: string;
-}
+// `ProjectInfo` now lives in `cli/skills/project-info.ts` (shared with the future `skills get setup`
+// CLI command); re-exported here so existing `cli/ai` importers (e.g. `cli/ai/mcp/`) keep compiling.
+export type { ProjectInfo } from '../skills/project-info.ts';
