+++ code/core/src/cli/skills/content/setup-prompts/monorepo-optimized-tests-relaxed-limits-no-story-deletion.ts
@@ -1,19 +1,20 @@
 import { dedent } from 'ts-dedent';
 
-import type { ProjectInfo, SetupInstructionsContext } from '../types.ts';
-import { getDocsMarkdownUrl } from '../utils/docs-markdown-url.ts';
-import { ext } from '../utils/ext.ts';
-import { listDOD, listRules, listSteps } from '../utils/markdown.ts';
+import type { ProjectInfo } from '../../project-info.ts';
+import { getDocsMarkdownUrl } from '../setup-utils/docs-markdown-url.ts';
+import { ext } from '../setup-utils/ext.ts';
+import { listDOD, listRules, listSteps } from '../setup-utils/markdown.ts';
+import type { SetupInstructionsContext } from './types.ts';
 import {
   buildPortalStep,
   buildSharedPreviewStep,
   cleanupStep,
-  discoveryStepStrict,
+  discoveryStepRelaxed,
   interactionPlayStep,
   monorepoStep,
   mswStep,
-  verifyStep,
-  writeStoriesStep,
+  verifyWithAllowedFailureStep,
+  writeStoriesWithAllowedFailuresStep,
 } from './partials/steps.ts';
 import {
   batchTestsRule,
@@ -23,15 +24,15 @@ import {
   noPolishRule,
   packageManagerRule,
   preferSharedFixesRule,
-  readBudgetRule,
+  readBudgetRuleRelaxed,
   toolsVsShellRule,
 } from './partials/rules.ts';
 import {
   cssCheckDOD,
   sharedPreviewDOD,
-  storyTagsV1DOD,
-  typeCheckPassesStrictDOD,
-  vitestPassesStrictDOD,
+  storyTagsV2DOD,
+  typeCheckPassesWhenExpectedDOD,
+  vitestPassesWhenExpectedDOD,
 } from './partials/dod.ts';
 
 export function instructions(projectInfo: ProjectInfo): string {
@@ -61,8 +62,8 @@ export function instructions(projectInfo: ProjectInfo): string {
     ${listRules([
       toolsVsShellRule(ctx),
       nodeModuleReadsRule(ctx),
-      monorepoRule(ctx),
-      readBudgetRule(ctx),
+      monorepoRule(projectInfo),
+      readBudgetRuleRelaxed(ctx),
       editOverWriteRule(ctx),
       batchTestsRule(ctx),
       packageManagerRule(ctx),
@@ -74,26 +75,26 @@ export function instructions(projectInfo: ProjectInfo): string {
 
     ${listSteps(
       [
-        discoveryStepStrict(projectInfo, ctx),
+        discoveryStepRelaxed(projectInfo, ctx),
         monorepoStep(projectInfo, ctx),
         buildSharedPreviewStep(projectInfo, ctx),
         buildPortalStep(projectInfo, ctx),
         mswStep(projectInfo, ctx),
-        writeStoriesStep(projectInfo, ctx),
+        writeStoriesWithAllowedFailuresStep(projectInfo, ctx),
         interactionPlayStep(projectInfo, ctx),
-        verifyStep(projectInfo, ctx),
+        verifyWithAllowedFailureStep(projectInfo, ctx),
         cleanupStep(projectInfo, ctx),
       ],
       { level: 3 }
     )}
 
     ## Done when
-    
+
     ${listDOD([
       cssCheckDOD(ctx),
-      storyTagsV1DOD(ctx),
-      vitestPassesStrictDOD(ctx),
-      typeCheckPassesStrictDOD(ctx),
+      storyTagsV2DOD(ctx),
+      vitestPassesWhenExpectedDOD(ctx),
+      typeCheckPassesWhenExpectedDOD(ctx),
       sharedPreviewDOD(ctx),
     ])}
