+++ code/core/src/cli/skills/content/setup-prompts/partials/types.ts
