+++ code/core/src/cli/skills/content/setup-utils/ext.ts
