+++ code/core/src/cli/skills/content/setup-utils/type-import-source.ts
@@ -1,4 +1,4 @@
-import type { ProjectInfo } from '../types.ts';
+import type { ProjectInfo } from '../../project-info.ts';
 
 export function getTypeImportSource(projectInfo: ProjectInfo): string {
   return projectInfo.framework || projectInfo.rendererPackage || '@storybook/react';
