+++ scripts/eval/lib/utils.ts
@@ -6,7 +6,7 @@ import { x } from 'tinyexec';
 import {
   DEFAULT_PROMPT_NAME,
   PROMPT_NAMES,
-} from '../../../code/core/src/cli/ai/setup-prompts/index.ts';
+} from '../../../code/core/src/cli/skills/content/setup-prompts/index.ts';
 import { getAiSetupPrompt } from '../../../code/core/src/shared/utils/ai-prompts.ts';
 
 export interface Logger {
